requirements.txt
scikit-learn
pandas
pyyaml
numpy
joblib
pytest  # for testing